fix frame cache replacement eviction
